Updates to sender and getting tests passing
